app/requirements.txt
flask
pyaes
samsungctl
wakeonlan
python-miio
